# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantum Lounge Social API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  password: string;
}

export interface AuthStatus {
  isOwner: boolean;
  loungeName?: string | null;
}

export interface Room {
  id: number;
  name: string;
  theme: string;
  frequency: number;
  capacity: number;
  guestCount: number;
  isOpen: boolean;
  createdAt: string;
}

export interface CreateRoomBody {
  name: string;
  theme: string;
  frequency: number;
  capacity: number;
  isOpen?: boolean;
}

export interface UpdateRoomBody {
  name?: string;
  theme?: string;
  frequency?: number;
  capacity?: number;
  isOpen?: boolean;
}

export interface Guest {
  id: number;
  name: string;
  vibe: string;
  /** basic | charged | quantum | transcended */
  energyLevel: string;
  roomId?: number | null;
  roomName?: string | null;
  /** active | checked_out */
  status: string;
  feePaid: number;
  checkedInAt: string;
  checkedOutAt?: string | null;
}

export interface CreateGuestBody {
  name: string;
  vibe: string;
  roomId?: number | null;
}

export interface UpdateGuestBody {
  status?: string;
  roomId?: number | null;
  vibe?: string;
  energyLevel?: string;
}

export interface Message {
  id: number;
  senderName: string;
  /** Quantum-encrypted message content */
  content: string;
  roomId?: number | null;
  roomName?: string | null;
  isEncrypted: boolean;
  /** Unique quantum hash for message verification */
  quantumSignature: string;
  sentAt: string;
}

export interface SendMessageBody {
  senderName: string;
  content: string;
  roomId?: number | null;
}

export interface EnergyUpgrade {
  id: number;
  /** e.g. Quantum Boost, Portal Ignition, Cosmic Transcendence */
  name: string;
  /** charged | quantum | transcended */
  level: string;
  description: string;
  price: number;
  color: string;
  createdAt: string;
}

export interface CreateEnergyUpgradeBody {
  name: string;
  level: string;
  description: string;
  price: number;
  color: string;
}

export interface GuestUpgrade {
  id: number;
  guestId: number;
  guestName: string;
  upgradeId: number;
  upgradeName: string;
  level: string;
  amountPaid: number;
  purchasedAt: string;
}

export interface PurchaseUpgradeBody {
  guestId: number;
  upgradeId: number;
}

export interface TeleportBody {
  guestId: number;
  fromRoomId?: number | null;
  toRoomId: number;
}

export interface TeleportResult {
  success: boolean;
  guest: Guest;
  message: string;
  /** Energy shift value from the teleport */
  quantumShift: number;
}

export interface TeleportEvent {
  id: number;
  guestId: number;
  guestName: string;
  fromRoomId?: number | null;
  fromRoomName?: string | null;
  toRoomId: number;
  toRoomName: string;
  quantumShift: number;
  teleportedAt: string;
}

export interface Transaction {
  id: number;
  /** house_fee | energy_upgrade | cashout */
  type: string;
  amount: number;
  description: string;
  guestId?: number | null;
  guestName?: string | null;
  /** completed | pending */
  status: string;
  createdAt: string;
}

export interface CashoutBody {
  amount: number;
  notes?: string | null;
}

export interface LoungeSettings {
  loungeName: string;
  tagline: string;
  houseFee: number;
  ownerName: string;
  theme: string;
  isOpen: boolean;
}

export interface UpdateSettingsBody {
  loungeName?: string;
  tagline?: string;
  houseFee?: number;
  ownerName?: string;
  theme?: string;
  isOpen?: boolean;
}

export interface DashboardStats {
  totalRevenue: number;
  availableBalance: number;
  totalCashedOut: number;
  activeGuests: number;
  totalGuests: number;
  openRooms: number;
  totalRooms: number;
  totalMessages: number;
  totalTeleports: number;
  recentTransactions: Transaction[];
}

export type ListMessagesParams = {
  roomId?: number;
};